tests/wpt/tests/css/css-borders/tentative/parsing/corner-shape-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Computed values of 'corner-shape'</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shaping">
<meta name="assert" content="This test checks that the computed value of 'corner-shape' is correct.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="target"></div>

<script>
test_computed_value("corner-top-left-shape", "round");
test_computed_value("corner-top-left-shape", "scoop");
test_computed_value("corner-top-left-shape", "superellipse(5)");
test_computed_value("corner-top-left-shape", "superellipse(0.2)");
test_computed_value("corner-top-left-shape", "superellipse(0)", "notch");
test_computed_value("corner-top-left-shape", "superellipse(infinity)", "straight");
test_computed_value("corner-top-left-shape", "superellipse(2)", "round");
test_computed_value("corner-top-left-shape", "superellipse(1)", "bevel");
test_computed_value("corner-top-left-shape", "superellipse(4)", "squircle");
test_computed_value("corner-top-left-shape", "superellipse( .5)", "scoop");
test_computed_value("corner-top-right-shape", "round");
test_computed_value("corner-top-right-shape", "superellipse(5)");
test_computed_value("corner-bottom-right-shape", "scoop");
test_computed_value("corner-bottom-left-shape", "superellipse(5)");
test_computed_value("corner-shape", "superellipse(5) round");
test_computed_value("corner-shape", "round");
test_computed_value("corner-shape", "bevel superellipse(0.1) round squircle");
test_computed_value("corner-shape", "superellipse(0.1) superellipse(3) superellipse(7) superellipse(0.1)");
test_computed_value("corner-shape", "round round round round", "round");
test_computed_value("corner-shape", "round scoop");
test_computed_value("corner-shape", "round scoop round scoop", "round scoop");
test_computed_value("corner-shape", "bevel superellipse(2)", "bevel round");
test_computed_value("corner-shape", "superellipse(0.5) superellipse(3) straight", "scoop superellipse(3) straight");
test_computed_value("corner-shape", "superellipse(0.5) superellipse(3) superellipse(1)", "scoop superellipse(3) bevel");
test_computed_value("corner-shape", "bevel superellipse(2) squircle     round", "bevel round squircle");
test_computed_value("corner-shape", "superellipse(0.5) superellipse(3) superellipse(1) superellipse(infinity)", "scoop superellipse(3) bevel straight");

</script>
